Add ability to post a story
The the submit story button now works when a marker has been
placed and all fields have values in them. The query sent to the
database by clicking the submit button successfully inserts a
new story into the database. the success callback in the client
then refreshes the page to show the new marker affixed to the
map.