Fix which elements are displayed when clicking different buttons
Clicking a new marker after looking into the sub-stories or
substory submission form of another marker will now properly hide
the substory list and substory submission form. Also adjusted the
color of text that appears on dark backgrounds to white for
readability. need to adjust the positioning of one piece of text
to center.