Clean up code according to style guide
Cleaned up code in component files and server index to match
airbnb style guide guidelines. Left in console and alert messages
that are still useful for now, and commented out unused variables
that may or may not be needed in the future. Will delete if end
up being worthless. Installed prop-types to help match guidelines.