Render substory list
Clicking the 'show additional files' button under a files will
now correctly render and list the names of all associated
subfiles. CSS stylings were also applied to the list
to give it the same tan background as the other form areas.
Next steps are to make it when a substory is clicked in the
substory list, a page detailing that substory is rendered,
with an included button to go back to the substory list.
Also need to remove uneeded elements and buttons when others
are rendered.
Additional changes:
- updated webpack to autoresolve .jsx files and not require
.jsx to be specified in their import statements
- added a 'clickable' class to the stylesheet that changes
the cursor to a pointer when hovering over clickable elements
- changed a method to static in substory.jsx and removed
the constructor