Add ability to go back to sub story menu after clickin one
The substory form now fully renders with all stylings in the
standard story form. Additionally, a 'go back' button was placed
at the top of the substory render area that allows users to go
back to the substory list menu and click on a new substory to
view.

Other minor changes:
- fixed spacing between words in 'my weird town' title at top
of page
- refactored the case number at the top of each story and
substory to correctly be 10000 + id number

Bugs still needing fixing:
- 'render addendum files' button still does not disappear when
it should
- 'go back' button appears as part of the substory form, but
should instead be above it
- case number in substory creation form currently reads as NaN

Future features to work on:
- general page styling improvements (particularily the substory)
- ability to organize substories by date created and number
of likes
- add ability to like a story or substory
- display number of likes for a story/ substory