Make like button on clickable once
clicking the like button now deactivates the like button and
reactivates the dislike button. clicking the dislike button
deactivates the dislike button and activates the like button.
Currently this is tied to the like button element as a whole,
so it creates a bug where clicking like on one substory
locks the ability to like any substory.