Add authentication
Add ability to login and logout using Auth0. Currently logging
in does not change the functionality of the website in any way.

Additionally:
-Added ability to access env files in react by installing
dotenv-webpack, along with the necessary code in the
webpack.config
-Updated react and react-dom versions so application can now
take advantage of hooks.

Note for future update:
-Refactor app so that interactive elements are given their own
separate files to be more compliant with MVC concepts.