Show which, if any, user is currently logged in
Logging in to a user account will now display the name of that
user at the top of the screen. If no user is logged in, a
message stating 'no current user' will be displayed instead.
Additionally, began refactoring the database schema to
accomodate data for likes and posts by users, though currently
incomplete.