Refactor like and dislike
Refactored the like and dislike and dislike buttons for the
substories to be housed in their own separate jsx files, and
additionally changed them to activate based on whether or not
a user is logged in. Trying to like or dislike a story while
not logged in will display a prompt asking the user to log in.

Additional changes:
-changed the database schema to have tables for users, story
likes, and substory likes
-added an 'app map' file that visually shows how the application
is set up file-structure-wise