Refactor components
Refactored several components to no longer be class based if
it was not needed for them to be class based. Moved functions
handling interactive elements of those components to helper files.

Additional Changes:
-updated appmap
Bug Fixes:
-Fixed bug where submitting a substory would not work properly.
Cause of this bug was a prop accidently being initialized as a
state variable
Known bugs that need addressing:
-Clicking on the map to add map markers is currently not
functioning properly
-Adding stories and substories to the database relays the message
that the submission was succesful, but neither are actually added
to the database