Fix story submit bug
Fixed a bug in which the subitting a story or substory would
display a message client side that the submission was succesful,
but the submitted entry would not be added to the database.
Issue was a missing field (for number of likes) in the db
query statement. Currently now succesfully adds the submission
to the database, but a purposefully failed submission still
displays a false-positive on submission.