Refactor substory form handling
Refactored several components so that the substory form
textarea values are held in state rather than grabbing from the
DOM. Also added a 'handleUserError' function to app.jsx that
will handle user errors like trying to submit a form with an
empty input field. This function was then passed down to
SubStorySubmit.