Set up city selection
Set up the components to let users select a starting location.
The submit button for picking a city will make a server call
that will then make a call to the google geocoding API, but
currently the call is not complete, meaning the user is unable
to get past the initial splash page.